backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
pydantic
python-dotenv


